REFACTOR: Moved landing hero component to common
