+++ src/components/common/LandingHero.vue
